change display of values from float (e.g. 0.34) to int (e.g. 34)

+++ Backend/Entities/GraphNodes/AudioFeaturesFilters/FilterDurationMsNode.cs
@@ -8,6 +8,6 @@ namespace Backend.Entities.GraphNodes.AudioFeaturesFilters
 {
     public class FilterDurationMsNode : FilterRangeNode
     {
-        protected override double? GetValue(Track t) => t.DurationMs;
+        protected override int? GetValue(Track t) => t.DurationMs;
     }
 }
