+++ Backend/Entities/GraphNodes/FilterYearNode.cs
@@ -4,7 +4,7 @@ namespace Backend.Entities.GraphNodes
 {
     public class FilterYearNode : FilterRangeNode
     {
-        protected override double? GetValue(Track t) => t.Album.ReleaseYear;
+        protected override int? GetValue(Track t) => t.Album.ReleaseYear;
         public override bool RequiresAlbums => true;
     }
 }
